requirements.txt
PyQt6==6.7.1
paho-mqtt==2.1.0
pydantic==2.8.2
pyqtgraph==0.13.7
tomli==2.0.1
